Sensor and Holder Rework
This makes the API the source of truth for the sensors and their configs.
Loading the devices (index) page will fetch available sensors and holders
from the API to consume.

In practice, the "config" tab has _never_ been used and thus the power
to create and edit sensors and holders now lies with the developers.