Metadata Rework: No more typescript types parsed
The idea until now was to parse the metadata.xlsx into a .yaml file for
loading the data and into a d.ts file to create type-safety when
building the components for it.

This is however an approach that mixes data and data structure.

We now only parse the metadata.xlsx into a yaml file for the data and
have a fixed types file on hand. The structure of how a parameter and a
profile are defined will not change unless we change the system, so the
types that go along it are not required to be code-generated.

This is the first step in removing the metadata responsibility from the
client. We want to shift it to ICOapi to use interchangeable files, and
the client can't easily access the filesystem.